+++ basic_system/src/system_implementation/flat_storage_model/preimage_cache.rs
@@ -16,7 +16,7 @@ use zk_ee::{
 
 use super::cost_constants::PREIMAGE_CACHE_GET_NATIVE_COST;
 use super::*;
-use crate::system_implementation::flat_storage_model::cost_constants::blake2s_native_cost;
+use crypto::blake2s::blake2s_native_cost;
 
 /// Query ID for requesting preimage data from the flat storage system
 pub const FLAT_STORAGE_GENERIC_PREIMAGE_QUERY_ID: u32 =
